+++ etna/metrics/utils.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import Callable
 from typing import Dict
 from typing import List
@@ -37,24 +38,89 @@ def compute_metrics(
     return metrics_values
 
 
+def mean_agg():
+    """Mean for pandas agg."""
+
+    def func(x: pd.Series):
+        with warnings.catch_warnings():
+            # this helps to prevent warning in case of all nans
+            warnings.filterwarnings(
+                message="Mean of empty slice",
+                action="ignore",
+            )
+            return np.nanmean(a=x.values)
+
+    func.__name__ = "mean"
+    return func
+
+
+def median_agg():
+    """Median for pandas agg."""
+
+    def func(x: pd.Series):
+        with warnings.catch_warnings():
+            # this helps to prevent warning in case of all nans
+            warnings.filterwarnings(
+                message="All-NaN slice encountered",
+                action="ignore",
+            )
+            return np.nanmedian(a=x.values)
+
+    func.__name__ = "median"
+    return func
+
+
+def std_agg():
+    """Std for pandas agg."""
+
+    def func(x: pd.Series):
+        with warnings.catch_warnings():
+            # this helps to prevent warning in case of all nans
+            warnings.filterwarnings(
+                message="Degrees of freedom <= 0",
+                action="ignore",
+            )
+            return np.nanstd(a=x.values)
+
+    func.__name__ = "std"
+    return func
+
+
+def notna_size_agg():
+    """Size of not-na elements for pandas agg."""
+
+    def func(x: pd.Series):
+        return len(x) - pd.isna(x.values).sum()
+
+    func.__name__ = "notna_size"
+    return func
+
+
 def percentile(n: int):
     """Percentile for pandas agg."""
 
-    def percentile_(x):
-        return np.nanpercentile(a=x.values, q=n)
+    def func(x: pd.Series):
+        with warnings.catch_warnings():
+            # this helps to prevent warning in case of all nans
+            warnings.filterwarnings(
+                message="All-NaN slice encountered",
+                action="ignore",
+            )
+            return np.nanpercentile(a=x.values, q=n)
 
-    percentile_.__name__ = f"percentile_{n}"
-    return percentile_
+    func.__name__ = f"percentile_{n}"
+    return func
 
 
 MetricAggregationStatistics = Literal[
-    "median", "mean", "std", "percentile_5", "percentile_25", "percentile_75", "percentile_95"
+    "median", "mean", "std", "notna_size", "percentile_5", "percentile_25", "percentile_75", "percentile_95"
 ]
 
 METRICS_AGGREGATION_MAP: Dict[MetricAggregationStatistics, Union[str, Callable]] = {
-    "median": "median",
-    "mean": "mean",
-    "std": "std",
+    "median": mean_agg(),
+    "mean": median_agg(),
+    "std": std_agg(),
+    "notna_size": notna_size_agg(),
     "percentile_5": percentile(5),
     "percentile_25": percentile(25),
     "percentile_75": percentile(75),
@@ -62,7 +128,7 @@ METRICS_AGGREGATION_MAP: Dict[MetricAggregationStatistics, Union[str, Callable]]
 }
 
 
-def aggregate_metrics_df(metrics_df: pd.DataFrame) -> Dict[str, float]:
+def aggregate_metrics_df(metrics_df: pd.DataFrame) -> Dict[str, Optional[float]]:
     """Aggregate metrics in :py:meth:`log_backtest_metrics` method.
 
     Parameters
@@ -74,7 +140,7 @@ def aggregate_metrics_df(metrics_df: pd.DataFrame) -> Dict[str, float]:
     if "fold_number" in metrics_df.columns:
         metrics_dict = (
             metrics_df.groupby("segment")
-            .mean()
+            .mean(numeric_only=False)
             .reset_index()
             .drop(["segment", "fold_number"], axis=1)
             .apply(list(METRICS_AGGREGATION_MAP.values()))
@@ -85,10 +151,11 @@ def aggregate_metrics_df(metrics_df: pd.DataFrame) -> Dict[str, float]:
     else:
         metrics_dict = metrics_df.drop(["segment"], axis=1).apply(list(METRICS_AGGREGATION_MAP.values())).to_dict()
 
-    metrics_dict_wide = {
-        f"{metrics_key}_{statistics_key}": value
-        for metrics_key, values in metrics_dict.items()
-        for statistics_key, value in values.items()
-    }
+    cur_dict = {}
+    for metrics_key, values in metrics_dict.items():
+        for statistics_key, value in values.items():
+            new_key = f"{metrics_key}_{statistics_key}"
+            new_value = value if not pd.isna(value) else None
+            cur_dict[new_key] = new_value
 
-    return metrics_dict_wide
+    return cur_dict